server/requirements.txt
absl-py==0.6.1
aniso8601==6.0.0
astor==0.7.1
astroid==2.1.0
attrs==18.2.0
autobahn==18.11.2
Automat==0.7.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
constantly==15.1.0
cycler==0.10.0
Cython==0.29.6
dask==1.0.0
decorator==4.3.0
ExifRead==2.1.2
ez-setup==0.9
fastzbarlight==0.0.14
Flask==1.0.2
Flask-Cors==3.0.7
Flask-HTTPAuth==3.2.4
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.7
Flask-WTF==0.14.2
future==0.17.1
gast==0.2.0
go-vncdriver==0.4.19
gpsphoto==2.2.2
grpcio==1.16.1
h5py==2.8.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mysql-connector-python==8.0.15
mysqlclient==1.4.2.post1
networkx==2.2
passlib==1.7.1
piexif==1.1.2
protobuf==3.6.1
pyglet==1.3.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.2.2
PyMySQL==0.9.3
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyWavelets==1.0.1
PyYAML==3.13
requests==2.10.0
reverse-geocoder==1.5.1
rope==0.11.0
six==1.11.0
termcolor==1.1.0
toolz==0.9.0
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.1.0
ujson==1.35
urllib3==1.24.1
uWSGI==2.0.18
websocket-client==0.54.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
zope.interface==4.6.0
Flask-Mail==0.9.1 
blinker==1.4